refactor(collab-server): split blob GC policy from blob GC mechanism
Two reviewers flagged blob-gc.ts against the ~400-line rule, and unlike the
other size findings I dismissed today this one is mine to fix: I CREATED this
file in #2794 already over the limit (464 lines), rather than inheriting an
overrun. Adding the boot sweep took it to 501.

Split at a seam worth having beyond the line count:

  blob-gc.ts        (262) - WHAT to collect: reference scanning, planning,
                            deletion. Pure, no scheduling, no environment.
  blob-gc-worker.ts (273) - WHEN and WHETHER to run: the worker lifecycle,
                            environment parsing, and the safety bounds.

Everything in the second file is policy an operator can misconfigure, and it is
exactly where both #2804 defects lived: a schedule that never fired at boot, and
a grace floor set at the destructive value. Keeping them together makes the
dangerous surface one file rather than a third of a bigger one.

No behaviour change. API surface snapshot unchanged (the split re-exports both
modules from index.ts), 177/177 across 36 files, typecheck and source-text gates
clean.
